transition to EL equations
as opposed to minimizing action (which - turns out - is physically wrong, we need to find stationary points instead of minima), we try minimizing deviation from the EL equations.

i am committing this for posterity but the current gradient tape implementation is expensive and heavy on retracing.
next step -- try discretized kernels for differential operators.